refactor: Extract task dismissible logic into a separate widget
- Create a new `TaskDismissible` widget to encapsulate dismissible task behavior
- Move existing dismissible task implementation from `TaskListView` to the new widget
- Improve code organization and reusability of task dismissal functionality
- Maintain existing interaction and visual design for task dismissal
